Add ProjectConfig with env loader and update scripts
