+++ src/config/__init__.py
@@ -1,0 +1,5 @@
+"""Configuraciones base del proyecto."""
+
+from .base import ProjectConfig
+
+__all__ = ["ProjectConfig"]
