+++ src/pipelines/ml/step_7_0_train_models.py
@@ -81,20 +81,22 @@ def configure_gpu(use_gpu=USE_GPU, memory_limit=GPU_MEMORY_LIMIT):
 has_gpu = configure_gpu()
 
 # Importar configuraciones centralizadas
-from src.core.config.settings import settings
+from src.config.base import ProjectConfig
 
-PROJECT_ROOT = settings.project_root
-MODELS_DIR = settings.models_dir
-TRAINING_DIR = settings.training_dir
-RESULTS_DIR = settings.results_dir
-IMG_CHARTS_DIR = settings.img_charts_dir
-DATE_COL = settings.date_col
-LOCAL_REFINEMENT_DAYS = settings.local_refinement_days
-TRAIN_TEST_SPLIT_RATIO = settings.train_test_split_ratio
-FORECAST_HORIZON_1MONTH = settings.forecast_horizon_1month
-FORECAST_HORIZON_3MONTHS = settings.forecast_horizon_3months
-RANDOM_SEED = settings.random_seed
-ensure_directories = settings.ensure_dirs
+config = ProjectConfig.from_env()
+
+PROJECT_ROOT = config.project_root
+MODELS_DIR = config.models_dir
+TRAINING_DIR = config.training_dir
+RESULTS_DIR = config.results_dir
+IMG_CHARTS_DIR = config.img_charts_dir
+DATE_COL = config.date_col
+LOCAL_REFINEMENT_DAYS = config.local_refinement_days
+TRAIN_TEST_SPLIT_RATIO = config.train_test_split_ratio
+FORECAST_HORIZON_1MONTH = config.forecast_horizon_1month
+FORECAST_HORIZON_3MONTHS = config.forecast_horizon_3months
+RANDOM_SEED = config.random_seed
+ensure_directories = config.ensure_dirs
 
 # Importar funciones de visualización
 from utils.plots import (
